1) Merged both ExcelHandler and ExcelWrite classes code into ExcelHandler.
